Numerous simplification changes
- Use the `build:` property in docker-compose.yaml so that the images
will automatically build, rather than having a separate manual process

- Simplify service and hostnames in docker-compose.yaml. That's changing
"the_caller" to just "caller", etc.  Note that underscores are a poor
choice in docker service names, becuase the underscore isn't a valid
character for hostname purposes.

- Remove docker-compose dependancies.  There's no reason why kamailio
has to start before the callee service.  Kamailio and callee shoudl be
started before the caller service, but since the actual call is
initiated manually, there's no reason to require other containers to be
running before this containner.

- Set `/usr/local/sbin/kamailio` as the kamailio container entrypoint.
This should probably be done in the docker file.

- Change volume mounts from specific files to directories.  When
mounting files, editing the file and restarting the service will throw
an error.  Mounting the directory allows the file to be edited and the
service to be restarted, rather than having to delete and create a new
container.

- Renamed directory "dockerfiles/kamailio 5.7.5" to
"dockerfiles/kamailio_5.7.5" as spaces in file names are a pain to
work with.

- Add `#!KAMAILIO` to top of kamailio config, as it's standard

- Removed `version:` from docker file as it's depricated.

- Load `tm` module before `sl` module.

- Change `memdbg` and `memlog` levels.  They were unnecessarily high.

- Set logging level to 2 (INFO) in docker-compose command for kamailio.

- Set color output in docker-compose command for kamailio.

- Set memory levels in docker-compose command for kamailio (best practices).

- .gitignore .swp files (for vim)